+++ src/Mock/EventDispatcherInterfaceMock.php
@@ -10,7 +10,6 @@ namespace PrecisionSoft\Symfony\Phpunit\Mock;
 
 use Closure;
 use Mockery\MockInterface;
-use PrecisionSoft\Symfony\Phpunit\Container\MockContainer;
 use PrecisionSoft\Symfony\Phpunit\Contract\MockDtoInterface;
 use PrecisionSoft\Symfony\Phpunit\MockDto;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
@@ -29,7 +28,7 @@ class EventDispatcherInterfaceMock implements MockDtoInterface
 
     public static function getOnCreate(): Closure
     {
-        return static function (MockInterface $mockInterface, MockContainer $mockContainer): void {
+        return static function (MockInterface $mockInterface): void {
             $mockInterface->shouldReceive('dispatch')
                 ->byDefault()
                 ->andReturnUsing(static function (object $event, ?string $eventName = null): object {