+++ src/Mock/SluggerInterfaceMock.php
@@ -10,7 +10,6 @@ namespace PrecisionSoft\Symfony\Phpunit\Mock;
 
 use Closure;
 use Mockery\MockInterface;
-use PrecisionSoft\Symfony\Phpunit\Container\MockContainer;
 use PrecisionSoft\Symfony\Phpunit\Contract\MockDtoInterface;
 use PrecisionSoft\Symfony\Phpunit\MockDto;
 use Symfony\Component\String\AbstractUnicodeString;
@@ -31,7 +30,7 @@ class SluggerInterfaceMock implements MockDtoInterface
 
     public static function getOnCreate(): Closure
     {
-        return static function (MockInterface $mockInterface, MockContainer $mockContainer): void {
+        return static function (MockInterface $mockInterface): void {
             $mockInterface->shouldReceive('slug')
                 ->byDefault()
                 ->andReturnUsing(static function (string $string): AbstractUnicodeString {